Move Round into its own module
Problem: Round type is defined in Block.hs module, but it's useful
in other places that shouldn't depend on Block.

Solution:
1. Create Data.Algorand.Round with the Round type.
2. Change `spLastRound` to be `Round` because that's a round.
3. Same for 2 fields in `Transaction`.
